removing nltk dependency from the split_sentences function
I reworked the split_sentences function at line 202 so that it no longer depends on the nltk library. With this function I am able to recreate the example you have in your readme file. In the testing I've done it seems to work just fine, but there might be some edge cases I haven't discovered yet.