Removed "Philippine" and "Manila" from stopwords
Some news articles about Philippines and Manila are important. There is no reason to have them as stop words on the main repository. If you want to put specific countries and cities on your stop words list, then please do so on a separate branch and not on the main repository.